# Calculate confidence interval for benchmark measurements