--- a/scripts/ci/e2e_benchmark.py
+++ b/scripts/ci/e2e_benchmark.py
@@ -1,4 +1,4 @@
-pimport requests
+import requests
 import sys
 import random
 from collections import defaultdict