--- a/scripts/ci/e2e_benchmark.py
+++ b/scripts/ci/e2e_benchmark.py
@@ -41,6 +41,10 @@ class BenchmarkRunner:
                     code = response.json()['code']
                     if code == 'NoSegment' or code == 'NoMatch':
                         continue
+                elif benchmark_name == 'route':
+                    code = response.json()['code']
+                    if code == 'NoRoute':
+                        continue
                 raise Exception(f"Error: {response.status_code} {response.text}")
             times.append((end_time - start_time) * 1000) # convert to ms