--- a/scripts/ci/e2e_benchmark.py
+++ b/scripts/ci/e2e_benchmark.py
@@ -130,7 +130,7 @@ def main():
 
     print('Shape: ', all_times.shape)
 
-    total_time, total_ci, total_min = calculate_confidence_interval(np.sum(all_times, axis=1))
+    total_time, total_ci, total_best = calculate_confidence_interval(np.sum(all_times, axis=1))
     min_time, min_ci, _ = calculate_confidence_interval(np.min(all_times, axis=1))
     mean_time, mean_ci, _ = calculate_confidence_interval(np.mean(all_times, axis=1))
     median_time, median_ci, _ = calculate_confidence_interval(np.median(all_times, axis=1))
@@ -138,7 +138,7 @@ def main():
     perc_99_time, perc_99_ci, _ = calculate_confidence_interval(np.percentile(all_times, 99, axis=1))
     max_time, max_ci, _ = calculate_confidence_interval(np.max(all_times, axis=1))
 
-    print(f'Total: {total_time:.2f}ms ± {total_ci:.2f}ms. Min: {total_min:.2f}ms')
+    print(f'Total: {total_time:.2f}ms ± {total_ci:.2f}ms. Best: {total_best:.2f}ms')
     print(f"Min time: {min_time:.2f}ms ± {min_ci:.2f}ms")
     print(f"Mean time: {mean_time:.2f}ms ± {mean_ci:.2f}ms")
     print(f"Median time: {median_time:.2f}ms ± {median_ci:.2f}ms")