--- a/scripts/ci/e2e_benchmark.py
+++ b/scripts/ci/e2e_benchmark.py
@@ -7,8 +7,6 @@ import csv
 import numpy as np
 import time
 import argparse
-from scipy import stats
-import logging
 
 
 class BenchmarkRunner: