--- a/scripts/ci/e2e_benchmark.py
+++ b/scripts/ci/e2e_benchmark.py
@@ -1,5 +1,4 @@
 import requests
-import sys
 import random
 from collections import defaultdict
 import os