This PR:
- adds confidence intervals to some of benchmark results(i.e. instead of `avg: 42ms` we now show `avg: 42 +- 2ms`, but to achieve that each benchmark now runs 5 times on CI
- improves benchmark scripts, so it becomes easier to run them locally
- some of benchmark cases are removed to make benchmarks faster (e.g. there is no need to run map matching with 4 different radiuses, it is enough to have 2)
- added new ops/sec metric to each bencmark
- fix bug in `bench` utility when we used huge (up to 5k points) traces as input for map matching API which caused significant slow downs of benchmarks

It still doesn't make benchmarks very stable and I still quite often see huge "jumps" between master and PR, but it is a step to direction of making it more stable(in general IMO we should eventually have some separate machine for benchmarks to avoid "noisy neighbours" on GH actions). 




<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1104.12<br/>plain u32: 1094.34<br/>aliased double: 956.011<br/>plain double: 947.599 | aliased u32: 1100.2<br/>plain u32: 1112.57<br/>aliased double: 954.001<br/>plain double: 951.872 |
| e2e_match_ch | Ops: 46.27 ± 0.12 ops/s. Best: 46.12 ops/s<br/>Total: 2831.65ms ± 7.31ms. Best: 2821.57ms<br/>Min time: 2.23ms ± 0.01ms<br/>Mean time: 21.61ms ± 0.06ms<br/>Median time: 16.20ms ± 0.07ms<br/>95th percentile: 69.59ms ± 0.26ms<br/>99th percentile: 84.18ms ± 0.15ms<br/>Max time: 97.50ms ± 0.50ms | Ops: 47.32 ± 0.14 ops/s. Best: 47.18 ops/s<br/>Total: 2768.90ms ± 8.19ms. Best: 2750.26ms<br/>Min time: 2.24ms ± 0.04ms<br/>Mean time: 21.13ms ± 0.06ms<br/>Median time: 15.53ms ± 0.24ms<br/>95th percentile: 66.75ms ± 0.26ms<br/>99th percentile: 78.91ms ± 0.31ms<br/>Max time: 93.24ms ± 0.43ms |
| e2e_match_mld | Ops: 62.03 ± 0.09 ops/s. Best: 61.88 ops/s<br/>Total: 2111.71ms ± 3.15ms. Best: 2105.63ms<br/>Min time: 1.86ms ± 0.04ms<br/>Mean time: 16.12ms ± 0.02ms<br/>Median time: 8.72ms ± 0.05ms<br/>95th percentile: 53.51ms ± 0.15ms<br/>99th percentile: 61.85ms ± 0.17ms<br/>Max time: 72.31ms ± 0.37ms | Ops: 66.83 ± 0.15 ops/s. Best: 66.49 ops/s<br/>Total: 1960.26ms ± 4.77ms. Best: 1954.97ms<br/>Min time: 1.83ms ± 0.03ms<br/>Mean time: 14.96ms ± 0.04ms<br/>Median time: 7.96ms ± 0.12ms<br/>95th percentile: 49.21ms ± 0.20ms<br/>99th percentile: 57.21ms ± 0.25ms<br/>Max time: 66.24ms ± 0.38ms |
| e2e_nearest_ch | Ops: 792.25 ± 4.10 ops/s. Best: 783.83 ops/s<br/>Total: 1262.22ms ± 6.92ms. Best: 1251.52ms<br/>Min time: 1.07ms ± 0.01ms<br/>Mean time: 1.26ms ± 0.01ms<br/>Median time: 1.17ms ± 0.01ms<br/>95th percentile: 1.67ms ± 0.01ms<br/>99th percentile: 1.73ms ± 0.01ms<br/>Max time: 4.35ms ± 2.61ms | Ops: 743.93 ± 9.35 ops/s. Best: 727.35 ops/s<br/>Total: 1344.51ms ± 18.36ms. Best: 1319.59ms<br/>Min time: 1.07ms ± 0.01ms<br/>Mean time: 1.34ms ± 0.02ms<br/>Median time: 1.20ms ± 0.01ms<br/>95th percentile: 1.81ms ± 0.07ms<br/>99th percentile: 2.97ms ± 0.20ms<br/>Max time: 6.29ms ± 1.35ms |
| e2e_nearest_mld | Ops: 802.40 ± 7.51 ops/s. Best: 783.80 ops/s<br/>Total: 1246.33ms ± 12.16ms. Best: 1234.80ms<br/>Min time: 1.06ms ± 0.01ms<br/>Mean time: 1.25ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.65ms ± 0.01ms<br/>99th percentile: 1.74ms ± 0.08ms<br/>Max time: 4.48ms ± 2.22ms | Ops: 744.44 ± 27.13 ops/s. Best: 699.31 ops/s<br/>Total: 1347.32ms ± 49.70ms. Best: 1289.08ms<br/>Min time: 1.06ms ± 0.00ms<br/>Mean time: 1.35ms ± 0.05ms<br/>Median time: 1.19ms ± 0.01ms<br/>95th percentile: 1.93ms ± 0.25ms<br/>99th percentile: 2.90ms ± 0.67ms<br/>Max time: 7.11ms ± 1.48ms |
| e2e_route_ch | Ops: 329.26 ± 3.47 ops/s. Best: 322.92 ops/s<br/>Total: 3035.96ms ± 31.12ms. Best: 2998.25ms<br/>Min time: 1.32ms ± 0.01ms<br/>Mean time: 3.04ms ± 0.03ms<br/>Median time: 3.07ms ± 0.02ms<br/>95th percentile: 4.01ms ± 0.06ms<br/>99th percentile: 4.42ms ± 0.08ms<br/>Max time: 7.05ms ± 2.25ms | Ops: 316.14 ± 11.74 ops/s. Best: 300.28 ops/s<br/>Total: 3173.61ms ± 118.26ms. Best: 3014.25ms<br/>Min time: 1.31ms ± 0.03ms<br/>Mean time: 3.17ms ± 0.12ms<br/>Median time: 3.13ms ± 0.07ms<br/>95th percentile: 4.47ms ± 0.37ms<br/>99th percentile: 6.12ms ± 0.98ms<br/>Max time: 9.07ms ± 1.11ms |
| e2e_route_mld | Ops: 287.98 ± 1.78 ops/s. Best: 284.37 ops/s<br/>Total: 3473.56ms ± 22.72ms. Best: 3446.35ms<br/>Min time: 1.29ms ± 0.01ms<br/>Mean time: 3.47ms ± 0.02ms<br/>Median time: 3.51ms ± 0.04ms<br/>95th percentile: 4.72ms ± 0.03ms<br/>99th percentile: 5.20ms ± 0.08ms<br/>Max time: 7.33ms ± 1.58ms | Ops: 269.30 ± 4.16 ops/s. Best: 263.78 ops/s<br/>Total: 3713.24ms ± 57.66ms. Best: 3600.81ms<br/>Min time: 1.31ms ± 0.01ms<br/>Mean time: 3.71ms ± 0.07ms<br/>Median time: 3.72ms ± 0.05ms<br/>95th percentile: 5.16ms ± 0.15ms<br/>99th percentile: 6.45ms ± 0.67ms<br/>Max time: 8.98ms ± 0.66ms |
| e2e_table_ch | Ops: 305.07 ± 1.96 ops/s. Best: 301.96 ops/s<br/>Total: 3277.56ms ± 21.07ms. Best: 3247.34ms<br/>Min time: 1.68ms ± 0.02ms<br/>Mean time: 3.28ms ± 0.02ms<br/>Median time: 3.27ms ± 0.02ms<br/>95th percentile: 4.53ms ± 0.05ms<br/>99th percentile: 4.88ms ± 0.10ms<br/>Max time: 8.12ms ± 2.40ms | Ops: 299.80 ± 2.08 ops/s. Best: 295.20 ops/s<br/>Total: 3334.58ms ± 23.24ms. Best: 3315.45ms<br/>Min time: 1.69ms ± 0.02ms<br/>Mean time: 3.34ms ± 0.02ms<br/>Median time: 3.32ms ± 0.02ms<br/>95th percentile: 4.64ms ± 0.06ms<br/>99th percentile: 5.20ms ± 0.22ms<br/>Max time: 8.90ms ± 1.54ms |
| e2e_table_mld | Ops: 108.28 ± 1.59 ops/s. Best: 104.95 ops/s<br/>Total: 9237.90ms ± 145.04ms. Best: 9080.94ms<br/>Min time: 3.70ms ± 0.07ms<br/>Mean time: 9.24ms ± 0.14ms<br/>Median time: 9.20ms ± 0.13ms<br/>95th percentile: 14.10ms ± 0.25ms<br/>99th percentile: 15.16ms ± 0.26ms<br/>Max time: 20.47ms ± 4.07ms | Ops: 106.82 ± 0.74 ops/s. Best: 105.67 ops/s<br/>Total: 9360.04ms ± 64.87ms. Best: 9266.90ms<br/>Min time: 3.76ms ± 0.04ms<br/>Mean time: 9.36ms ± 0.07ms<br/>Median time: 9.34ms ± 0.07ms<br/>95th percentile: 14.24ms ± 0.11ms<br/>99th percentile: 15.46ms ± 0.20ms<br/>Max time: 20.24ms ± 1.69ms |
| e2e_trip_ch | Ops: 96.16 ± 0.55 ops/s. Best: 95.04 ops/s<br/>Total: 10400.64ms ± 59.69ms. Best: 10305.98ms<br/>Min time: 1.56ms ± 0.13ms<br/>Mean time: 10.40ms ± 0.06ms<br/>Median time: 9.88ms ± 0.08ms<br/>95th percentile: 18.34ms ± 0.10ms<br/>99th percentile: 20.00ms ± 0.13ms<br/>Max time: 26.57ms ± 4.10ms | Ops: 92.23 ± 1.31 ops/s. Best: 90.05 ops/s<br/>Total: 10839.53ms ± 158.68ms. Best: 10669.79ms<br/>Min time: 1.67ms ± 0.19ms<br/>Mean time: 10.85ms ± 0.16ms<br/>Median time: 10.37ms ± 0.16ms<br/>95th percentile: 18.89ms ± 0.18ms<br/>99th percentile: 20.62ms ± 0.32ms<br/>Max time: 23.83ms ± 1.28ms |
| e2e_trip_mld | Ops: 56.88 ± 1.32 ops/s. Best: 54.81 ops/s<br/>Total: 17609.93ms ± 430.88ms. Best: 16922.75ms<br/>Min time: 1.82ms ± 0.18ms<br/>Mean time: 17.59ms ± 0.43ms<br/>Median time: 17.20ms ± 0.40ms<br/>95th percentile: 28.72ms ± 0.72ms<br/>99th percentile: 30.56ms ± 0.64ms<br/>Max time: 33.54ms ± 1.84ms | Ops: 55.40 ± 0.47 ops/s. Best: 54.57 ops/s<br/>Total: 18046.40ms ± 162.29ms. Best: 17749.95ms<br/>Min time: 1.72ms ± 0.22ms<br/>Mean time: 18.05ms ± 0.18ms<br/>Median time: 17.65ms ± 0.14ms<br/>95th percentile: 29.29ms ± 0.22ms<br/>99th percentile: 31.54ms ± 0.20ms<br/>Max time: 39.64ms ± 3.73ms |
| json-render | String: 7.45622ms<br/>Stringstream: 10.6755ms<br/>Vector: 6.85354ms | String: 6.64681ms<br/>Stringstream: 9.45432ms<br/>Vector: 6.97276ms |
| match_ch | Default radius: <br/>4.64322ms/req at 82 coordinate<br/>0.0566246ms/coordinate<br/>Radius 5m: <br/>4.62759ms/req at 82 coordinate<br/>0.056434ms/coordinate<br/>Radius 10m: <br/>16.0754ms/req at 82 coordinate<br/>0.196041ms/coordinate<br/>Radius 15m: <br/>39.239ms/req at 82 coordinate<br/>0.478525ms/coordinate<br/>Radius 30m: <br/>336.981ms/req at 82 coordinate<br/>4.10952ms/coordinate | Default radius: <br/>4.47327ms/req at 82 coordinate<br/>0.0545521ms/coordinate<br/>Radius 10m: <br/>15.2426ms/req at 82 coordinate<br/>0.185885ms/coordinate |
| match_mld | Default radius: <br/>3.18365ms/req at 82 coordinate<br/>0.038825ms/coordinate<br/>Radius 5m: <br/>3.13708ms/req at 82 coordinate<br/>0.0382571ms/coordinate<br/>Radius 10m: <br/>11.7563ms/req at 82 coordinate<br/>0.14337ms/coordinate<br/>Radius 15m: <br/>29.9239ms/req at 82 coordinate<br/>0.364925ms/coordinate<br/>Radius 30m: <br/>346.859ms/req at 82 coordinate<br/>4.22999ms/coordinate | Default radius: <br/>2.79943ms/req at 82 coordinate<br/>0.0341394ms/coordinate<br/>Radius 10m: <br/>10.3753ms/req at 82 coordinate<br/>0.126528ms/coordinate |
| osrm_contract | Time: 96.58s Peak RAM: 196.04MB | Time: 95.67s Peak RAM: 185.66MB |
| osrm_customize | Time: 1.37s Peak RAM: 116.67MB | Time: 1.31s Peak RAM: 115.26MB |
| osrm_extract | Time: 12.09s Peak RAM: 403.17MB | Time: 12.07s Peak RAM: 413.81MB |
| osrm_partition | Time: 2.02s Peak RAM: 135.80MB | Time: 2.16s Peak RAM: 148.92MB |
| packedvector | random write:<br/>std::vector 9881.38 ms<br/>util::packed_vector 81985.6 ms<br/>slowdown: 8.29698<br/>random read:<br/>std::vector 8432.27 ms<br/>util::packed_vector 30989.4 ms<br/>slowdown: 3.6751 | random write:<br/>std::vector 9850.49 ms<br/>util::packed_vector 81826.4 ms<br/>slowdown: 8.30684<br/>random read:<br/>std::vector 8571.45 ms<br/>util::packed_vector 32923.7 ms<br/>slowdown: 3.84109 |
| random_match_ch | 1000 matches, default radius<br/>total: 7054.47ms<br/>avg: 7.05ms<br/>min: 0.00ms<br/>max: 501.08ms<br/>p99: 109.76ms<br/><br/>1000 matches, radius=10<br/>total: 35260.86ms<br/>avg: 35.26ms<br/>min: 0.00ms<br/>max: 1910.36ms<br/>p99: 1869.08ms<br/><br/>1000 matches, radius=20<br/>total: 67980.72ms<br/>avg: 67.98ms<br/>min: 0.00ms<br/>max: 9270.73ms<br/>p99: 1170.25ms | 500 matches, default radius<br/>ops: 227.43 ± 0.93 ops/s. best: 225.15ops/s.<br/>total: 250.64 ± 1.03ms. best: 249.75ms.<br/>avg: 4.40 ± 0.02ms<br/>min: 0.15 ± 0.01ms<br/>max: 22.95 ± 0.06ms<br/>p99: 22.95 ± 0.06ms<br/><br/>500 matches, radius=10<br/>ops: 67.70 ± 0.06 ops/s. best: 67.64ops/s.<br/>total: 945.40 ± 0.76ms. best: 943.67ms.<br/>avg: 14.77 ± 0.01ms<br/>min: 0.15 ± 0.00ms<br/>max: 214.26 ± 0.89ms<br/>p99: 214.26 ± 0.89ms<br/><br/>500 matches, radius=20<br/>ops: 15.76 ± 0.07 ops/s. best: 15.67ops/s.<br/>total: 4124.58 ± 18.13ms. best: 4096.79ms.<br/>avg: 63.46 ± 0.28ms<br/>min: 0.30 ± 0.00ms<br/>max: 1104.51 ± 7.63ms<br/>p99: 1104.51 ± 7.63ms |
| random_match_mld | 1000 matches, default radius<br/>total: 5873.04ms<br/>avg: 5.87ms<br/>min: 0.00ms<br/>max: 441.54ms<br/>p99: 82.59ms<br/><br/>1000 matches, radius=10<br/>total: 30491.07ms<br/>avg: 30.49ms<br/>min: 0.00ms<br/>max: 1772.97ms<br/>p99: 1734.61ms<br/><br/>1000 matches, radius=20<br/>total: 58897.53ms<br/>avg: 58.90ms<br/>min: 0.00ms<br/>max: 7754.15ms<br/>p99: 911.83ms | 500 matches, default radius<br/>ops: 342.24 ± 2.08 ops/s. best: 337.21ops/s.<br/>total: 166.56 ± 1.02ms. best: 165.57ms.<br/>avg: 2.92 ± 0.02ms<br/>min: 0.12 ± 0.00ms<br/>max: 17.98 ± 0.07ms<br/>p99: 17.98 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 120.45 ± 0.16 ops/s. best: 120.12ops/s.<br/>total: 531.36 ± 0.69ms. best: 530.67ms.<br/>avg: 8.30 ± 0.01ms<br/>min: 0.15 ± 0.00ms<br/>max: 101.85 ± 0.20ms<br/>p99: 101.85 ± 0.20ms<br/><br/>500 matches, radius=20<br/>ops: 23.33 ± 0.02 ops/s. best: 23.31ops/s.<br/>total: 2785.64 ± 2.37ms. best: 2782.07ms.<br/>avg: 42.86 ± 0.04ms<br/>min: 0.20 ± 0.00ms<br/>max: 546.55 ± 1.13ms<br/>p99: 546.55 ± 1.13ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>total: 456.69ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.24ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 600.24ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.17ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 759.15ms<br/>avg: 0.08ms<br/>min: 0.03ms<br/>max: 0.22ms<br/>p99: 0.14ms | 10000 nearest, number_of_results=1<br/>ops: 21975.01 ± 190.44 ops/s. best: 21486.23ops/s.<br/>total: 455.12 ± 4.01ms. best: 452.06ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.20 ± 0.01ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 16629.80 ± 27.18 ops/s. best: 16568.88ops/s.<br/>total: 601.33 ± 1.00ms. best: 600.40ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13105.54 ± 10.10 ops/s. best: 13091.43ops/s.<br/>total: 763.04 ± 0.59ms. best: 761.93ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.14 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>total: 459.13ms<br/>avg: 0.05ms<br/>min: 0.01ms<br/>max: 0.21ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 600.80ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.16ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 753.47ms<br/>avg: 0.08ms<br/>min: 0.03ms<br/>max: 0.21ms<br/>p99: 0.14ms | 10000 nearest, number_of_results=1<br/>ops: 21859.96 ± 29.44 ops/s. best: 21810.65ops/s.<br/>total: 457.46 ± 0.62ms. best: 456.52ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.20 ± 0.01ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 16385.43 ± 188.99 ops/s. best: 16078.06ops/s.<br/>total: 610.41 ± 7.08ms. best: 601.97ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.23 ± 0.07ms<br/>p99: 0.13 ± 0.01ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13000.73 ± 29.56 ops/s. best: 12946.38ops/s.<br/>total: 769.19 ± 1.75ms. best: 766.58ms.<br/>avg: 0.08 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.24 ± 0.05ms<br/>p99: 0.14 ± 0.00ms |
| random_route_ch | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 20080.28ms<br/>avg: 2.01ms<br/>min: 0.13ms<br/>max: 4.17ms<br/>p99: 2.97ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 9374.27ms<br/>avg: 0.94ms<br/>min: 0.06ms<br/>max: 3.18ms<br/>p99: 1.59ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 17306.33ms<br/>avg: 1.73ms<br/>min: 0.07ms<br/>max: 5.23ms<br/>p99: 4.10ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 10431.25ms<br/>avg: 1.04ms<br/>min: 0.12ms<br/>max: 2.14ms<br/>p99: 1.50ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 4208.02ms<br/>avg: 0.42ms<br/>min: 0.05ms<br/>max: 0.77ms<br/>p99: 0.64ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 9176.90ms<br/>avg: 0.92ms<br/>min: 0.06ms<br/>max: 4.91ms<br/>p99: 2.36ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 648.92ms<br/>avg: 0.06ms<br/>min: 0.01ms<br/>max: 1.48ms<br/>p99: 0.64ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 666.56ms<br/>avg: 0.07ms<br/>min: 0.01ms<br/>max: 0.48ms<br/>p99: 0.39ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 901.19ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 2.14ms<br/>p99: 1.16ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 481.35 ± 0.43 ops/s. best: 480.53ops/s.<br/>total: 2044.26 ± 1.84ms. best: 2041.81ms.<br/>avg: 2.08 ± 0.00ms<br/>min: 0.36 ± 0.01ms<br/>max: 3.35 ± 0.13ms<br/>p99: 2.96 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 562.20 ± 1.17 ops/s. best: 560.58ops/s.<br/>total: 1778.73 ± 3.80ms. best: 1773.11ms.<br/>avg: 1.78 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.40 ± 0.11ms<br/>p99: 3.96 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 990.08 ± 6.09 ops/s. best: 979.22ops/s.<br/>total: 993.91 ± 6.13ms. best: 985.68ms.<br/>avg: 1.01 ± 0.01ms<br/>min: 0.26 ± 0.01ms<br/>max: 1.53 ± 0.06ms<br/>p99: 1.38 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1093.41 ± 3.11 ops/s. best: 1086.93ops/s.<br/>total: 914.58 ± 2.61ms. best: 911.49ms.<br/>avg: 0.91 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 2.96 ± 0.02ms<br/>p99: 2.32 ± 0.01ms |
| random_route_mld | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 39584.30ms<br/>avg: 3.96ms<br/>min: 0.11ms<br/>max: 12.88ms<br/>p99: 6.67ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 14248.83ms<br/>avg: 1.42ms<br/>min: 0.06ms<br/>max: 3.74ms<br/>p99: 2.52ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 41555.11ms<br/>avg: 4.16ms<br/>min: 0.06ms<br/>max: 13.81ms<br/>p99: 8.42ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 29677.57ms<br/>avg: 2.97ms<br/>min: 0.11ms<br/>max: 9.16ms<br/>p99: 5.19ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 9044.71ms<br/>avg: 0.90ms<br/>min: 0.05ms<br/>max: 2.80ms<br/>p99: 1.62ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 33009.41ms<br/>avg: 3.30ms<br/>min: 0.05ms<br/>max: 15.40ms<br/>p99: 6.52ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 835.55ms<br/>avg: 0.08ms<br/>min: 0.01ms<br/>max: 4.40ms<br/>p99: 1.56ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 901.20ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.20ms<br/>p99: 0.98ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 1714.47ms<br/>avg: 0.17ms<br/>min: 0.01ms<br/>max: 5.08ms<br/>p99: 3.53ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 234.67 ± 4.91 ops/s. best: 222.30ops/s.<br/>total: 4196.37 ± 91.99ms. best: 4126.09ms.<br/>avg: 4.26 ± 0.09ms<br/>min: 0.38 ± 0.02ms<br/>max: 9.56 ± 0.33ms<br/>p99: 7.10 ± 0.20ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 236.11 ± 0.70 ops/s. best: 235.02ops/s.<br/>total: 4235.33 ± 12.57ms. best: 4217.97ms.<br/>avg: 4.24 ± 0.01ms<br/>min: 0.06 ± 0.00ms<br/>max: 10.16 ± 1.42ms<br/>p99: 8.38 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 318.69 ± 2.61 ops/s. best: 313.88ops/s.<br/>total: 3087.87 ± 25.55ms. best: 3049.77ms.<br/>avg: 3.14 ± 0.03ms<br/>min: 0.30 ± 0.00ms<br/>max: 8.79 ± 1.91ms<br/>p99: 5.54 ± 0.21ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 297.43 ± 2.80 ops/s. best: 294.39ops/s.<br/>total: 3362.59 ± 31.75ms. best: 3292.65ms.<br/>avg: 3.36 ± 0.03ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.76 ± 0.55ms<br/>p99: 6.66 ± 0.16ms |
…